refactor portfolio and mktplace nav items to subcomponents
